+++ react/components/RichTextEditor/LinkInput.tsx
@@ -4,9 +4,11 @@ import { defineMessages, InjectedIntl, injectIntl } from 'react-intl'
 import { Button, IconLink, Input } from 'vtex.styleguide'
 
 import StyleButton from './StyleButton'
+import { useClickOutside } from './utils'
 
 interface Props {
-  onAdd: (link: string) => void
+  onAdd: (text: string, link: string) => void
+  getCurrentSelection: () => string | null
   intl: InjectedIntl
 }
 
@@ -15,28 +17,60 @@ const messages = defineMessages({
     defaultMessage: 'Add',
     id: 'admin/pages.admin.rich-text-editor.add-button',
   },
+  insertLink: {
+    id: 'admin/pages.admin.rich-text-editor.link.title',
+    defaultMessage: 'Insert link',
+  },
+  text: {
+    defaultMessage: 'Text',
+    id: 'admin/pages.admin.rich-text-editor.add-image.text-label',
+  },
 })
 
-const LinkInput = ({ onAdd, intl }: Props) => {
+const LinkInput = ({ onAdd, getCurrentSelection, intl }: Props) => {
+  const ref = React.useRef<HTMLDivElement>(null)
+
   const [isOpen, setIsOpen] = React.useState(false)
   const [link, setLink] = React.useState()
+  const [text, setText] = React.useState()
 
-  const handleAddImage = () => {
+  useClickOutside(ref, () => setIsOpen(false))
+
+  const handleAddLink = () => {
     setIsOpen(false)
-    return onAdd(link)
+    return onAdd(text, link)
+  }
+
+  const handleToggleDialog = () => {
+    if (isOpen) {
+      return setIsOpen(false)
+    }
+
+    setText(getCurrentSelection())
+    setIsOpen(true)
   }
 
   return (
-    <div className="relative">
+    <div className="relative" ref={ref}>
       <StyleButton
+        title={intl.formatMessage(messages.insertLink)}
         active={isOpen}
-        onToggle={() => setIsOpen(!isOpen)}
+        onToggle={handleToggleDialog}
         style={null}
         label={<IconLink />}
       />
 
       {isOpen && (
-        <div className="flex flex-column absolute pa5 bg-white b--solid b--muted-4 bw1 br2 w5">
+        <div className="flex flex-column absolute pa5 bg-white b--solid b--muted-4 bw1 br2 w5 z-1">
+          <div className="mb4">
+            <Input
+              label={intl.formatMessage(messages.text)}
+              value={text}
+              onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                setText(e.target.value)
+              }
+            />
+          </div>
           <div className="mb4">
             <Input
               label={'URL'}
@@ -45,7 +79,11 @@ const LinkInput = ({ onAdd, intl }: Props) => {
               }
             />
           </div>
-          <Button onClick={handleAddImage} size="small">
+          <Button
+            disabled={!text || !link}
+            onClick={handleAddLink}
+            size="small"
+          >
             {intl.formatMessage(messages.btn)}
           </Button>
         </div>