+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Changed
+
+- Improve rich text editor
+- Save page slug based on page title
+
 ## [4.12.0] - 2019-09-11
 
 ### Changed